Animate PITCHf/x

Pitch trajectories animated on a two-dimensional plot.

animateFX(data, point.alpha = 1/3, layers, strike = FALSE,
  color = "pitch_type_name", x_lim = c(-3.5, 3.5), y_lim = c(0, 7),
  flag = FALSE, interval = 0.01, ...)

Arguments

data
data frame with appropriately named PITCHf/x variables

point.alpha
ggplot2 alpha parameter

layers
list of ggplot2 layer modifications.

color
character. plot point color.

x_lim
vector(2 element). x axis range.

y_lim
vector(2 element). y axis range.

interval
time (in seconds) between plotting the pitch locations.

...
extra options passed onto geom commands

Value

Returns a series of objects of the class used by package ggplot2 to represent plots.

Details

animateFX plots a series of "snapshots" that represent pitch trajectories from the point of release until all of them reach home plate. The graphic takes on the viewpoint of the umpire; that is, the pitches are getting closer to the viewer with time. This is relected with the increase in size of the "balls" as the animation progresses.

Examples

data(pitches)
#generate animation and prompt default web browser to view the sequence of plots
# NOT RUN {
animation::saveHTML({ animateFX(pitches, layer = facet_grid(pitcher_name~stand)) })
# }